moved resumecontroller helper functions to Helper>ResumeController_helpers - cleaning up ResumeController, Removed embedding comparisong with DB skills - Fix in future implementation (vector retrieval)
